A tiny FindBugs session on effects.
git-svn-id: http://ardorlabs.svn.cvsdude.com/Ardor3Dv1/trunk@1175 82abaa13-a93f-4eff-8987-b3ae9fc18050

+++ ardor3d-effects/src/main/java/com/ardor3d/extension/terrain/ClipmapLevel.java
@@ -314,8 +314,8 @@ public class ClipmapLevel extends Mesh {
         } else {
             // indices of heightvalues we can use
             // for the second height to avoid cracks
-            float xLow = (x / vertexDistance);
-            float zLow = (z / vertexDistance);
+            float xLow = (x / (float) vertexDistance);
+            float zLow = (z / (float) vertexDistance);
             if (xLow < 0) {
                 xLow -= 1.0f;
             }
